Consolidate LCIA mapping scripts into single file
